requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
customtkinter==5.2.2
darkdetect==0.8.0
discord.py==2.4.0
frozenlist==1.4.1
idna==3.9
multidict==6.1.0
packaging==24.1
pillow==10.4.0
toml==0.10.2
yarl==1.11.1
